Refactor client and controller to remove duplicate code, and pass client settings to lower-level request functions
