+++ pyinaturalist/controllers/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa: F401
 from pyinaturalist.controllers.base import BaseController
-from pyinaturalist.controllers.observation import ObservationController
+from pyinaturalist.controllers.observations import ObservationController
